That update is a complete OpenGL integration with both Orbiters.
The API is almost clean and need minimum configurations for work any new effect.
Too is integrated with Orbiter.conf which makes easy to enable/disable it.

There is a separation from current drawings. I fix up any constructor with bUseOpenGL
globally to make Orbiter aware of OpenGL. 

Texture Manager inside the DrawingWidgetsEngine which means that the video drawback is 
reduced to minimum (the exception is on PocketFrog where OpenGL is destroed on every
animation) and too means that every effect should not register the textures for animation.