Adds some bullet proofing to some of the MythTV scripts.
Note: The patch to mythfilldatabase to ignore backend will be in a separate commit.

This avoids some of the problems with channel scanning on multicore machines.

This loads up todays listing data before telling the user that the backend is ready,
so that they can tune into LiveTV without 'unknown' program issues.

This checks for the existance of downloaded icons before trying to move them.

Also, once the --ignore-backend option is added to mythfilldatabase, FillDbAndFetchIcons.sh will use it to separate loading listings data from the potentially crash inducing command to force a reschedule.