Updates 50_lmce_keybindings patch so that it will work without mheg patches applied.
The New Zealand MHEG patches from the Ubuntu mythtv packages cause deadlocks with low bitrate HDTV streams; so those are disabled in the 00list, but the 50_lmce_keybindings patch (which ensures that F6,F7,F8 are not used by MythTV) needed to be updated so as to still apply cleanly. I've created a new dpatch 50_lmce_keybindings_no_mheg which we use instead. in 00list.