Syncs mythtv-0.20.2+fixes14472 to revision 14472 from upstream.. There were some missing fixes, including ones written specifically for LinuxMCE.
Also the logging was disabled. Two problems: first, this should have been done via dpatch and second, the reason it was disabled was because it caused deadlocks. But it only caused these because deadlock behaviour was requested in the old Ubuntu debian/rules file, this was an Ubuntu packaging error which was fixed there and we have that fix in our debian/rules now so we don't need the hack.