Fixes several bugs in MythTV_PlugIn::UpdateUpcomingRecordings().
This was apparently always broken. It tried to read the header within the loop that reads off the individual ProgramInfo entries, so it could only correctly read the first entry in the list.

There were also a couple off-by-one errors, but at least one of those may have crept in due to changes in the MythTV protocol since this was first written (it was at the end of the list) and the other initialized an unused variable. I named all the fields read in so that this will be easier to keep in sync in the future.