Keep track of the age of the remote assistance ports database record. Refresh the port allocation once a day.
How it works:

Remote assistance is started by the cron job along, not by a startup script.
The cron job runs once every minute. If it finds that the ports should be
refreshed, it refreshes them and re-reads the database.

If the case of a machine that was offline for more than a week,
SetupRemoteAssistance notices that the ports need to be refreshed, at boot,
and it gets new ports from the server before it starts any tunnels.

If the machine was offline for less than a week, but more than a day,
SetupRemoteAssistance will re-confirm the ports (it will get the same ones),
this way also telling the server that the machine didn't vanish into the void.

This fixes:
- stale port information (old information in the database)
- periodical port confirmation (so the server doesn't reallocate the ports)