Refs #879 - OSD Orbiter does not display after stopping non-pluto media device
Reintruducing the code that was removed in changeset r23325 with one
modification. Now the MD will only be turned on when the media that
was stopped was using Live AV.  This is needed to set any device
pipes back to the inputs of the Media Director when a Live AV device
is done playing.