Refs #879 - OSD Orbiter does not display after stopping non-pluto media device
The previous change (r23379) did not properly handle the case when an audio-only
LiveAV device was played, such as an OTARadio.  Under that circumstance the MD's
display would be turned on when it should not be.

This fix, like r23325, gets rid of the special code to turn on the MD.  It folds
that logic into the HandleOnOffs routine and properly handles the case that an
audio only device is being turned off.  In that case, the display will not be
turned On.

This change depends on #880 to not turn on the display for any media type other
than the pluto CD, Audio File, and Live Radio MediaTypes.