Fixes #746 - Follow me volume does not work properly
When viewing a non-pluto device through a capture card, volume controls
are now sent to the Media Director in the Entertainment Area where the
device is being viewed.

Previously the volume controls were always sent directly to the device -
or to its audio pipe.  Now this only happens when watching the device
via direct/live AV.